fix: PostgreSQL 호환성 해결 및 인증 로직 수정 (#4) - 쿠키 기반 refresh/logout 제거하고 JWT 로그인 응답을 JSON으로 통일 - 주요 인증 로직을 EmailAuthService로 이전 - WebSecurityConfig에서 JWT 필터 추가 및 쿠키 관련 설정 제거 - AuthController에 회원가입 API 추가 - PostgreSQL 호환성 문제 해결     - 'AUTO_INCREMENT'를 'GENERATED BY DEFAULT AS IDENTITY'로 변경     - 'ON UPDATE CURRENT_TIMESTAMP' 제거 - init_data.sql에 초기 데이터 삽입 스크립트 추가
