+++ src/main/java/sonmoeum/api/v1/auth/dto/response/LoginResponse.java
@@ -1,0 +1,8 @@
+package sonmoeum.api.v1.auth.dto.response;
+
+import sonmoeum.api.v1.users.dto.response.UserResponse;
+
+public record LoginResponse(
+    String accessToken,
+    UserResponse user
+) {}
