+++ src/main/java/sonmoeum/domain/auth/service/EmailAuthService.java
@@ -4,8 +4,11 @@ import java.util.List;
 
 import sonmoeum.api.v1.auth.dto.request.EmailLoginRequest;
 import sonmoeum.api.v1.auth.dto.request.EmailSignupRequest;
+import sonmoeum.api.v1.auth.dto.response.LoginResponse;
 import sonmoeum.api.v1.users.dto.request.CreateEmailUserRequest;
 import sonmoeum.api.v1.users.dto.response.UserResponse;
+import sonmoeum.common.security.jwt.JwtTokenProvider;
+import sonmoeum.common.security.service.CustomUserDetails;
 import sonmoeum.domain.users.service.UserCrudService;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -24,6 +27,7 @@ import lombok.RequiredArgsConstructor;
 public class EmailAuthService {
     private final UserCrudService userCrudService;
     private final AuthenticationManager authenticationManager;
+    private final JwtTokenProvider jwtTokenProvider;
 
     @Transactional
     public UserResponse signUp(EmailSignupRequest request) {
@@ -37,14 +41,19 @@ public class EmailAuthService {
         ));
     }
 
-    public void login(EmailLoginRequest request, HttpServletRequest httpRequest, HttpServletResponse httpResponse) {
-        UsernamePasswordAuthenticationToken token = 
-            new UsernamePasswordAuthenticationToken(request.email(), request.password());
-            
-        Authentication authentication = authenticationManager.authenticate(token);
-        
-        SecurityContext context = SecurityContextHolder.createEmptyContext();
-        context.setAuthentication(authentication);
-        SecurityContextHolder.setContext(context);
+    public LoginResponse login(EmailLoginRequest request) {
+        Authentication authentication = authenticationManager.authenticate(
+            new UsernamePasswordAuthenticationToken(request.email(), request.password())
+        );
+
+        CustomUserDetails userDetails = (CustomUserDetails) authentication.getPrincipal();
+        UserResponse userResponse = userCrudService.getUserById(userDetails.getUserId());
+
+        String accessToken = jwtTokenProvider.createAccessToken(
+            userDetails.getUsername(),
+            userDetails.getAuthorities()
+        );
+
+        return new LoginResponse(accessToken, userResponse);
     }
 }
